Refactor sys_call to windows_event_api
